all: move json and string functions into nwg_tools.
Functions: split_string, string_to_json, save_json.
