+++ common/nwg_tools.h
@@ -10,6 +10,15 @@
 #pragma once
 
 #include <string>
+#include <vector>
+#include <iostream>
+#include <iomanip>
+
+#include <gtkmm.h>
+
+#include <nlohmann/json.hpp>
+
+namespace ns = nlohmann;
 
 std::string get_config_dir(std::string);
 
@@ -19,5 +28,9 @@ std::string get_locale(void);
 
 std::string read_file_to_string(std::string);
 void save_string_to_file(std::string, std::string);
+std::vector<std::string> split_string(std::string, std::string);
+
+ns::json string_to_json(std::string);
+void save_json(ns::json, std::string);
 
 std::string get_output(std::string);
