+++ grid/grid.h
@@ -110,12 +110,9 @@ std::string get_cache_path(void);
 std::string get_pinned_path(void);
 void add_and_save_pinned(std::string);
 void remove_and_save_pinned(std::string);
-std::vector<std::string> split_string(std::string, std::string);
 std::vector<std::string> get_app_dirs(void);
 std::vector<std::string> list_entries(std::vector<std::string>);
 std::vector<std::string> desktop_entry(std::string, std::string);
-ns::json string_to_json(std::string);
-void save_json(ns::json, std::string);
 ns::json get_cache(std::string);
  std::vector<std::string> get_pinned(std::string);
 std::vector<CacheEntry> get_favourites(ns::json, int);
